+++ frontend/src/pages/MeetingInfo.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const MeetingInfo = () => {
+    return <div></div>;
+};
+
+export default MeetingInfo;
